It all works
